Decomm rpi8; Add new control plane node rpi11
rpi8 will be rebuilt later as a worker node.

rpi11 is the first Raspberry Pi 5 node with NVMe disk, joining control
plane as an experiment. If everything looks great in a few weeks, the
remaining 2 old control plane nodes will also be decomm'ed and replaced
by 2 new nodes with new hardware.

ip addresses are temporarily added to populate the etcd node IPs, which
is required to run the recover playbook.