Expand cnpg rbac; Add network policy
- Switch CloudNativePG to ClusterWide manifests
- Add a network policy to forbid egress to internet

Initially I put a lot of effort to restrict cnpg operator's access to
its own namespace because otherwise it would need permission to access
secrets in all namespaces, which I considered too wide.

This restriction means the operator is only able to manage database
clusters within its own namespace. And an application living in a
different naemspace would have to access the database via a ClusterIP
service across namespaces.

This leads to some problems, specifically in secret sharing, which
defeats the purpose of leveraging an operator to lift the burden of
database management for me.

Secret sharing is necessary for authentication, in either password auth
or mTLS, a CA would have to provision some secrets in both database's
namespace and the application's namespace.

Ultimately, as I deploy more applications with database backend,
databases become one of the most critical piece of the infrastructure.

As a temporary workaround, I add a network policy to kill the operator's
internet access -- the operator shouldn't be able to leak my secrets out
even if it's compromised. In the long run, I should deploy a policy
engine (OPA) to prevent unexpected secret access.